fix(nodes): Add name property to the nodes
