+++ nodes/bpm/check-property.html
@@ -3,6 +3,7 @@
         category: 'bpm',
         color: '#DDAA99',
         defaults: {
+            name: { value: "" },
             propertyToCheck: { value: "" },
             expectedValue: { value: "" },
             params: { 
@@ -34,6 +35,10 @@
 </script>
 
 <script type="text/x-red" data-template-name="check-property">
+    <div class="form-row">
+        <label for="node-input-name">Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
     <div class="form-row">
         <label for="node-input-propertyToCheck"><i class="fa fa-info"></i> Property</label>
         <input type="text" id="node-input-propertyToCheck" placeholder="Property">
