--- nodes/bpm/csv-chunk-reader.html
@@ -1,31 +1,0 @@
-<script type="text/javascript">
-    RED.nodes.registerType('csv-chunk-reader', {
-        category: 'bpm',
-        color: '#DDAA99',
-        defaults: {
-            name: { value: "" },
-            path: { value: "" },
-            linesPerChunk: { value: 1000 }
-        },
-        inputs: 1,
-        outputs: 1,
-        icon: 'font-awesome/fa-folder-open',
-        label: function() {
-            return this.name || 'csv-chunk-reader';
-        }
-    });
-</script>
-<script type="text/x-red" data-template-name="csv-chunk-reader">
-    <div class="form-row">
-        <label for="node-input-name">Name</label>
-        <input type="text" id="node-input-name" placeholder="Name">
-    </div>
-    <div class="form-row">
-        <label for="node-input-path">File Path</label>
-        <input type="text" id="node-input-path" placeholder="Path to CSV file">
-    </div>
-    <div class="form-row">
-        <label for="node-input-linesPerChunk">Lines Per Chunk</label>
-        <input type="number" id="node-input-linesPerChunk" placeholder="30000">
-    </div>
-</script>
