+++ nodes/bpm/find-object.html
@@ -3,6 +3,7 @@
         category: 'bpm',
         color: '#DDAA99',
         defaults: {
+            name: { value: "" },
             key: { value: "" },
             keyValue: { value: "" },
             params: { 
@@ -34,6 +35,10 @@
 </script>
 
 <script type="text/x-red" data-template-name="find-object">
+    <div class="form-row">
+        <label for="node-input-name">Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
     <div class="form-row">
         <label for="node-input-key"><i class="fa fa-key"></i> Key</label>
         <input type="text" id="node-input-key" placeholder="Key">
