+++ nodes/bpm/get-property-value.html
@@ -3,6 +3,7 @@
         category: 'bpm',
         color: '#DDAA99',
         defaults: {
+            name: { value: "" },
             propertyToGet: { value: "" },
             params: { 
                 value: {
@@ -32,6 +33,10 @@
 </script>
 
 <script type="text/x-red" data-template-name="get-property-value">
+    <div class="form-row">
+        <label for="node-input-name">Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
     <div class="form-row">
         <label for="node-input-propertyToGet"><i class="fa fa-font"></i> Property></label>
         <input type="text" id="node-input-propertyToGet" placeholder="Property">
