+++ nodes/bpm/group-by.html
@@ -26,6 +26,10 @@
 
 <script type="text/x-red" data-template-name="group-by">
     <div class="form-row">
+        <label for="node-input-name">Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
+    <div class="form-row">
         <label for="node-input-property"><i class="fa fa-tag"></i> Property</label>
         <input type="text" id="node-input-property" placeholder="Property to group by">
     </div>
