+++ nodes/bpm/group-by.js
@@ -2,6 +2,8 @@ module.exports = function (RED) {
     function GroupByPropertyNode(config) {
         RED.nodes.createNode(this, config);
         var node = this;
+        var groups = {};
+        var ac = 0;
         node.on("input", function (msg) {
             let property =
                 msg.req && msg.req.body && msg.req.body.property !== undefined
@@ -11,18 +13,19 @@ module.exports = function (RED) {
                 node.error("No property specified to group by.");
                 return;
             }
-
-            let groupedByProperty = msg.payload.reduce((acc, item) => {
-                let propKey = item[property];
-                if (!acc[propKey]) {
-                    acc[propKey] = [];
+            var payload = msg.payload;
+            if (payload && payload[property]) {
+                var key = payload[property];
+                if (!groups[key]) {
+                    groups[key] = [];
                 }
-                acc[propKey].push(item);
-                return acc;
-            }, {});
-            msg.items = Object.keys(groupedByProperty);
-            msg.payload = groupedByProperty;
-            node.send(msg);
+                groups[key].push(payload);
+                ac++;
+            }
+            if (ac === msg.len) {
+                msg.payload = groups;
+                node.send(msg);
+            }
         });
     }
     RED.nodes.registerType("group-by", GroupByPropertyNode);
