+++ nodes/bpm/has-activity.html
@@ -3,6 +3,7 @@
         category: 'bpm',
         color: '#DDAA99',
         defaults: {
+            name: { value: "" },
             conceptName: { value: "" },
             params: { 
                 value: {
@@ -32,6 +33,10 @@
 </script>
 
 <script type="text/x-red" data-template-name="has-activity">
+    <div class="form-row">
+        <label for="node-input-name">Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
     <div class="form-row">
         <label for="node-input-conceptName"><i class="fa fa-align-justify"></i> Concept Name</label>
         <input type="text" id="node-input-conceptName" placeholder="Concept Name">
