+++ nodes/bpm/last-timestamp.html
@@ -3,6 +3,7 @@
         category: 'bpm',
         color: '#DDAA99',
         defaults: {
+            name: { value: "" },
         },
         inputs: 1,
         outputs: 1,
@@ -26,4 +27,8 @@
 </script>
 
 <script type="text/x-red" data-template-name="last-timestamp">
+    <div class="form-row">
+        <label for="node-input-name">Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
 </script>
