+++ nodes/bpm/log-getter.html
@@ -3,6 +3,7 @@
         category: 'bpm',
         color: '#DDAA99',
         defaults: {
+            name: { value: "" },
             logUrl: { value: "" },
             params: {
                 value: {
@@ -32,6 +33,10 @@
 </script>
 
 <script type="text/x-red" data-template-name="log-getter">
+    <div class="form-row">
+        <label for="node-input-name">Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
     <div class="form-row">
         <label for="node-input-logUrl"><i class="fa fa-link"></i>Url</label>
         <input type="text" id="node-input-logUrl" placeholder="https://example/doc.pdf">
